Fixed PR issues, files were moved around and renamed.
