+++ Slide for Reddit/Comments/Extensions/CommentViewController+LinkCellViewDelegate+SubmissionMoreDelegate.swift
@@ -8,8 +8,39 @@
 
 import Foundation
 
-extension CommentViewController: LinkCellViewDelegate {
+extension CommentViewController: LinkCellViewDelegate, SubmissionMoreDelegate {
     // MARK: - Methods
+    // MARK: - Shared Methods
+    /// Saves a link cell.
+    @objc func save(_ cell: LinkCellView) {
+        do {
+            let state = !ActionStates.isSaved(s: cell.link!)
+            print(cell.link!.id)
+            try session?.setSave(state, name: (cell.link?.id)!, completion: { (result) in
+                if result.error != nil {
+                    print(result.error!)
+                }
+                DispatchQueue.main.async {
+                    BannerUtil.makeBanner(text: state ? "Saved" : "Unsaved", color: ColorUtil.accentColorForSub(sub: self.subreddit), seconds: 1, context: self)
+                }
+            })
+            ActionStates.setSaved(s: cell.link!, saved: !ActionStates.isSaved(s: cell.link!))
+            History.addSeen(s: cell.link!, skipDuplicates: true)
+            cell.refresh()
+            if parent is PagingCommentViewController {
+                (parent as! PagingCommentViewController).reloadCallback?()
+            }
+        } catch {
+            // Catch
+        }
+    }
+    
+    /// Hide a cell view.
+    func hide(_ cell: LinkCellView) {
+        // Add stuff here.
+    }
+    
+    // MARK: - LinkCellView Methods
     /// Sets an upvote from the user.
     @objc func upvote(_ cell: LinkCellView) {
         do {
@@ -45,29 +76,6 @@ extension CommentViewController: LinkCellViewDelegate {
         }
     }
     
-    /// Saves a link cell.
-    @objc func save(_ cell: LinkCellView) {
-            do {
-                let state = !ActionStates.isSaved(s: cell.link!)
-                print(cell.link!.id)
-                try session?.setSave(state, name: (cell.link?.id)!, completion: { (result) in
-                    if result.error != nil {
-                        print(result.error!)
-                    }
-                    DispatchQueue.main.async {
-                        BannerUtil.makeBanner(text: state ? "Saved" : "Unsaved", color: ColorUtil.accentColorForSub(sub: self.subreddit), seconds: 1, context: self)
-                    }
-                })
-                ActionStates.setSaved(s: cell.link!, saved: !ActionStates.isSaved(s: cell.link!))
-                History.addSeen(s: cell.link!, skipDuplicates: true)
-                cell.refresh()
-                if parent is PagingCommentViewController {
-                    (parent as! PagingCommentViewController).reloadCallback?()
-                }
-            } catch {
-            }
-        }
-    
     /// Displays more info and presents alert.
     func more(_ cell: LinkCellView) {
         if !offline {
@@ -80,11 +88,6 @@ extension CommentViewController: LinkCellViewDelegate {
         if !offline {
             VCPresenter.presentAlert(TapBehindModalViewController.init(rootViewController: ReplyViewController.init(submission: cell.link!, sub: cell.link!.subreddit, delegate: self)), parentVC: self)
         }
-    }
-    
-    /// Hide a cell view.
-    func hide(_ cell: LinkCellView) {
-        // Add stuff here.
     }
 
     /// Undefined
@@ -127,6 +130,52 @@ extension CommentViewController: LinkCellViewDelegate {
             (parent as! PagingCommentViewController).reloadCallback?()
         }
         cell.refresh()
+    }
+    
+    // MARK: - SubmissionMore Methods
+    /// Undefined
+    func showFilterMenu(_ cell: LinkCellView) {
+        // Not implemented
+    }
+    
+    /// Applies a filter to comments.
+    func applyFilters() {
+        if PostFilter.filter([submission!], previous: nil, baseSubreddit: "all").isEmpty {
+            self.navigationController?.popViewController(animated: true)
+        }
+    }
+    
+    /// Hides selected comments.
+    func hide(index: Int) {
+        if index >= 0 {
+            self.navigationController?.popViewController(animated: true)
+        }
+    }
+    
+    /// Follows / Subscribes to Reddit Lists.
+    func subscribe(link: RSubmission) {
+        let sub = link.subreddit
+        let alrController = UIAlertController.init(title: "Follow r/\(sub)", message: nil, preferredStyle: .alert)
+        if AccountController.isLoggedIn {
+            let somethingAction = UIAlertAction(title: "Subscribe", style: UIAlertAction.Style.default, handler: { (_: UIAlertAction!) in
+                Subscriptions.subscribe(sub, true, session: self.session!)
+                self.subChanged = true
+                BannerUtil.makeBanner(text: "Subscribed to r/\(sub)", color: ColorUtil.accentColorForSub(sub: sub), seconds: 3, context: self, top: true)
+            })
+            alrController.addAction(somethingAction)
+        }
+        
+        let somethingAction = UIAlertAction(title: "Casually subscribe", style: UIAlertAction.Style.default, handler: { (_: UIAlertAction!) in
+            Subscriptions.subscribe(sub, false, session: self.session!)
+            self.subChanged = true
+            BannerUtil.makeBanner(text: "r/\(sub) added to your subreddit list", color: ColorUtil.accentColorForSub(sub: sub), seconds: 3, context: self, top: true)
+        })
+        alrController.addAction(somethingAction)
+        
+        alrController.addCancelButton()
+        
+        alrController.modalPresentationStyle = .fullScreen
+        self.present(alrController, animated: true, completion: {})
     }
     
 }
