+++ Slide for Reddit/Comments/Extensions/CommentViewController+UIGestureRecognizerDelegate.swift
