+++ Slide for Reddit/Comments/Extensions/CommentViewController+UINavigationControllerDelegate.swift
