+++ Slide for Reddit/Comments/Extensions/CommentViewController+UISearchBarDelegate.swift
