+++ Slide for Reddit/Comments/Extensions/CommentViewController+UITableViewDataSource.swift
