+++ Slide for Reddit/Comments/Extensions/CommentViewController+UITableViewDelegate.swift
