+++ Slide for Reddit/Comments/Views/CommentDepthCell.swift
