--- Slide for Reddit/Extensions/CommentViewController+SubmissionMoreDelegate.swift
@@ -1,86 +1,0 @@
-//
-//  CommentViewController+SubmissionMoreDelegate.swift
-//  Slide for Reddit
-//
-//  Created by Josiah Agosto on 8/3/20.
-//  Copyright © 2020 Haptic Apps. All rights reserved.
-//
-
-import Foundation
-
-extension CommentViewController: SubmissionMoreDelegate {
-    // MARK: - Methods
-    /// Saves comments from selected cell.
-    func save(_ cell: LinkCellView) {
-        do {
-            let state = !ActionStates.isSaved(s: cell.link!)
-            print(cell.link!.id)
-            try session?.setSave(state, name: (cell.link?.id)!, completion: { (result) in
-                if result.error != nil {
-                    print(result.error!)
-                }
-                DispatchQueue.main.async {
-                    BannerUtil.makeBanner(text: state ? "Saved" : "Unsaved", color: ColorUtil.accentColorForSub(sub: self.subreddit), seconds: 1, context: self)
-                }
-            })
-            ActionStates.setSaved(s: cell.link!, saved: !ActionStates.isSaved(s: cell.link!))
-            History.addSeen(s: cell.link!, skipDuplicates: true)
-            cell.refresh()
-            if parent is PagingCommentViewController {
-                (parent as! PagingCommentViewController).reloadCallback?()
-            }
-        } catch {
-        }
-    }
-    
-    /// Undefined
-    func hide(_ cell: LinkCellView) {
-
-    }
-    
-    /// Undefined
-    func showFilterMenu(_ cell: LinkCellView) {
-        // Not implemented
-    }
-    
-    /// Applies a filter to comments.
-    func applyFilters() {
-        if PostFilter.filter([submission!], previous: nil, baseSubreddit: "all").isEmpty {
-            self.navigationController?.popViewController(animated: true)
-        }
-    }
-    
-    /// Hides selected comments.
-    func hide(index: Int) {
-        if index >= 0 {
-            self.navigationController?.popViewController(animated: true)
-        }
-    }
-    
-    /// Follows / Subscribes to Reddit Lists.
-    func subscribe(link: RSubmission) {
-        let sub = link.subreddit
-        let alrController = UIAlertController.init(title: "Follow r/\(sub)", message: nil, preferredStyle: .alert)
-        if AccountController.isLoggedIn {
-            let somethingAction = UIAlertAction(title: "Subscribe", style: UIAlertAction.Style.default, handler: { (_: UIAlertAction!) in
-                Subscriptions.subscribe(sub, true, session: self.session!)
-                self.subChanged = true
-                BannerUtil.makeBanner(text: "Subscribed to r/\(sub)", color: ColorUtil.accentColorForSub(sub: sub), seconds: 3, context: self, top: true)
-            })
-            alrController.addAction(somethingAction)
-        }
-        
-        let somethingAction = UIAlertAction(title: "Casually subscribe", style: UIAlertAction.Style.default, handler: { (_: UIAlertAction!) in
-            Subscriptions.subscribe(sub, false, session: self.session!)
-            self.subChanged = true
-            BannerUtil.makeBanner(text: "r/\(sub) added to your subreddit list", color: ColorUtil.accentColorForSub(sub: sub), seconds: 3, context: self, top: true)
-        })
-        alrController.addAction(somethingAction)
-        
-        alrController.addCancelButton()
-        
-        alrController.modalPresentationStyle = .fullScreen
-        self.present(alrController, animated: true, completion: {})
-    }
-    
-}
